Extract fleet multiplex connect test to keep stream suite under 1k.
Moves the interleaving coverage into its own module and stubs orchestrator
submit instead of reimplementing private enqueue/_runs wiring.
